+++ src/frontend/src/components/cards/card.tsx
@@ -13,17 +13,17 @@ export interface CardProps {
 const Card: React.FC<CardProps> = ({ id, title, href, icon: Icon }) => {
   return (
     <>
-      <div className="w-full bg-white rounded-2xl border border-gray-100 shadow-lg text-gray-700 py-6 px-4">
+      <div className="card-white text-gray-700">
         <div className="flex flex-col items-center justify-center">
           <div className="bg-gray-200 rounded-full p-4 text-primary">
             <Icon size={24} />
           </div>
-          <div className="w-full bg-white p-6 flex flex-col items-center gap-2">
+          <div className="w-full bg-white p-6 flex flex-col items-center">
             <p className="text-xl font-semibold text-gray-900">{title}</p>
-            <hr className="w-full py-4"/>
+            <div className="w-3/4 lg:w-full border-b border-gray-300 mt-4 mb-6"></div>
             <Button variant="outline">
               <Link href={href}>
-                Gerenciar
+                <a>Gerenciar</a>
               </Link>
             </Button>
           </div>
