+++ src/frontend/src/components/navigation/navbar.tsx
@@ -1,10 +1,19 @@
 import { HiAcademicCap, HiMenu, HiUserCircle, HiX } from "react-icons/hi";
-import { NavbarLinks } from "@utils/data";
 import { useState } from "react";
 import Link from "next/link";
 import Dropdown from "@components/elements/dropdown";
 
-export default function Navbar() {
+export interface NavbarLinkProps {
+  id: string;
+  title: string;
+  href: string;
+}
+
+export interface NavbarProps {
+  links: NavbarLinkProps[];
+}
+
+const Navbar: React.FC<NavbarProps> = ({ links }) => {
   const [opened, setOpened] = useState(false);
 
   return (
@@ -18,7 +27,7 @@ export default function Navbar() {
               <HiAcademicCap size={32} />
             </div>
             <div className="hidden lg:ml-6 lg:flex lg:items-center">
-              {NavbarLinks.map((item) => (
+              {links.map((item) => (
                 <Link href={item.href} key={item.id}>
                   <a className="h-full inline-flex items-center py-2 px-4">
                     {item.title}
@@ -64,7 +73,7 @@ export default function Navbar() {
       {opened && (
         <section className="lg:hidden bg-white shadow-lg py-8 text-gray-700 animate-fade-in">
           <div className="space-y-1 px-2 font-bold">
-            {NavbarLinks.map((item) => (
+            {links.map((item) => (
               <Link href={item.href} key={item.id}>
                 <a className="block px-4 py-2 text-lg">{item.title}</a>
               </Link>
@@ -75,14 +84,9 @@ export default function Navbar() {
               textSize="lg"
               options={[
                 {
-                  id: "blog",
-                  title: "Blog",
-                  href: "/",
-                },
-                {
-                  id: "ajuda",
-                  title: "Central de Ajuda",
-                  href: "/",
+                  id: "sair",
+                  title: "Sair",
+                  href: "/autenticacao/login",
                 },
               ]}
             >
@@ -95,3 +99,5 @@ export default function Navbar() {
     </nav>
   );
 }
+
+export default Navbar;
