+++ src/frontend/src/pages/admin/alunos.tsx
@@ -1,6 +1,6 @@
 import type { NextPage } from "next";
 import { useEffect, useState } from "react";
-import { HiOutlinePencilAlt, HiOutlineBriefcase } from "react-icons/hi";
+import { HiOutlinePencilAlt, HiOutlineTrash, HiOutlineUserGroup } from "react-icons/hi";
 /* templates */
 import { Meta } from "@templates/meta";
 import { Template } from "@templates/template";
@@ -8,37 +8,31 @@ import { Template } from "@templates/template";
 import { paginationComponentOptions } from "@utils/table";
 /* components */
 import DataTable from "react-data-table-component";
-import CadastrarProfessor from "@components/modal/form/professor";
+import CadastrarAluno from "@components/modal/form/aluno";
 import { EmptyTable } from "@components/empty/table";
 import Excluir from "@components/modal/delete";
-import moment from "moment";
+import Navbar from "@components/navigation/navbar";
+import { NavbarAdminLinks } from "@utils/data";
 
 const columns = [
+  {
+    id: "matricula",
+    name: "Matrícula",
+    selector: (row: any) => row.matricula,
+    sortable: true,
+    width: "20%",
+  },
   {
     id: "nome",
     name: "Nome",
     selector: (row: any) => row.nome,
     sortable: true,
+    width: "30%",
   },
   {
-    id: "nasc",
-    name: "Data Nasc.",
-    selector: (row: any) => {
-      return moment(row.dataCriacao).format("DD-MM-YYYY");
-    },
-    sortable: true,
-  },
-  {
-    id: "depto",
-    name: "Departamento",
-    selector: (row: any) => row.departamento,
-    sortable: true,
-    grow: 2,
-  },
-  {
-    id: "salario",
-    name: "Salário",
-    selector: (row: any) => row.salario,
+    id: "email",
+    name: "E-mail",
+    selector: (row: any) => row.email,
     sortable: true,
   },
   {
@@ -46,12 +40,19 @@ const columns = [
     sortable: false,
     right: true,
     grow: 0,
-    cell: () => (
+    cell: (props: any) => (
       <div className="flex gap-2">
+        <button
+          type="button"
+          className="text-danger p-1 hover:bg-gray-50 rounded-full transition duration-200"
+          onClick={() => deleteAluno(props.id)}
+        >
+          <HiOutlineTrash size={18} />
+        </button>
         <Excluir
-          title="Excluir Professor"
-          description="Tem certeza que deseja excluir esse professor?"
-          onClick={() => (console.log("Excluiu Professor!"))}
+          title="Excluir Aluno"
+          description="Tem certeza que deseja excluir esse aluno?"
+          onClick={() => (console.log("Excluiu Aluno!"))}
         />
         <button
           type="button"
@@ -63,42 +64,51 @@ const columns = [
     ),
   },
 ];
-/*
+
 const data = [
   {
     id: 1,
-    nome: "José Maria Silva",
-    nasc: "13/11/1960",
-    departamento: "Departamento de Ciência da Computação",
-    salario: "R$ 18.000,00",
+    matricula: "201920584",
+    nome: "Sandra Carvalho",
+    email: "sandra.carvalho@estudante.ufla.br",
   },
   {
     id: 2,
-    nome: "Maria José Costa",
-    nasc: "29/01/1961",
-    departamento: "Departamento de Ciência da Computação",
-    salario: "R$ 20.000,00",
+    matricula: "201820584",
+    nome: "Marcos Vilela",
+    email: "marcos.vilela@estudante.ufla.br",
+  },
+  {
+    id: 3,
+    matricula: "201929584",
+    nome: "Camila Rodrigues",
+    email: "camila.rodrigues@estudante.ufla.br",
   },
 ];
-*/
+
+const deleteAluno = async (id: number) => {
+  await fetch(`http://localhost:8080/api/aluno/excluir/${id}`, {
+     method: 'DELETE',
+  }).then((response) => {
+    console.log(response)
+    window.location.reload();
+  });
+};
+
 const Home: NextPage = () => {
   const [showCadastrar, setShowCadastrar] = useState(false);
-  const [professores, setProfessores] = useState([]);
+  const [alunos, setAlunos] = useState([]);
 
   useEffect(() => {
-    fetch("http://localhost:8080/api/professor/listar")
+    fetch('http://localhost:8080/api/aluno/listar')
       .then((response) => response.json())
       .then((data) => {
-        data.map( (professor: any) => {
-          professor.departamento = professor.departamento.nome;
-        })
-        setProfessores(data);
+        setAlunos(data);
       })
       .catch((err) => {
-        console.log(err.message);
+        console.log(err.message)
       });
   }, []);
-
 
   return (
     <Template
@@ -111,10 +121,11 @@ const Home: NextPage = () => {
         />
       }
     >
+      <Navbar links={NavbarAdminLinks} />
       <div className="container py-16">
         <div className="flex justify-between items-center">
-          <h2 className="text-gray-700">Professores</h2>
-          <CadastrarProfessor
+          <h2 className="text-gray-700">Alunos</h2>
+          <CadastrarAluno
             show={showCadastrar}
             setShow={setShowCadastrar}
           />
@@ -123,12 +134,12 @@ const Home: NextPage = () => {
         <div className="mt-8 overflow-x-auto animate-fade-in-up text-gray-700">
           <DataTable
             columns={columns}
-            data={professores}
+            data={alunos}
             pagination
             paginationComponentOptions={paginationComponentOptions}
             highlightOnHover
             pointerOnHover
-            noDataComponent={<EmptyTable title="Não há professores cadastrados :(" description="Cadastre um professor no botão Cadastrar!" icon={HiOutlineBriefcase} />}
+            noDataComponent={<EmptyTable title="Não há alunos cadastrados :(" description="Cadastre um aluno no botão Cadastrar!" icon={HiOutlineUserGroup} />}
           />
         </div>
       </div>
