+++ src/frontend/src/pages/admin/cursos.tsx
@@ -1,6 +1,6 @@
 import type { NextPage } from "next";
-import { useState } from "react";
-import { HiOutlinePencilAlt, HiOutlineBookOpen } from "react-icons/hi";
+import { useEffect, useState } from "react";
+import { HiOutlinePencilAlt, HiFolderOpen } from "react-icons/hi";
 /* templates */
 import { Meta } from "@templates/meta";
 import { Template } from "@templates/template";
@@ -8,23 +8,19 @@ import { Template } from "@templates/template";
 import { paginationComponentOptions } from "@utils/table";
 /* components */
 import DataTable from "react-data-table-component";
-import CadastrarDisciplina from "@components/modal/form/disciplina";
+import CadastrarCurso from "@components/modal/form/curso";
 import { EmptyTable } from "@components/empty/table";
 import Excluir from "@components/modal/delete";
+import Navbar from "@components/navigation/navbar";
+import { NavbarAdminLinks } from "@utils/data";
 
 const columns = [
-  {
-    id: "codigo",
-    name: "Código",
-    selector: (row: any) => row.codigo,
-    sortable: true,
-    width: "10%",
-  },
   {
     id: "nome",
     name: "Nome",
     selector: (row: any) => row.nome,
     sortable: true,
+    width: "40%",
   },
   {
     id: "depto",
@@ -33,22 +29,23 @@ const columns = [
     sortable: true,
   },
   {
-    id: "professor",
-    name: "Professor",
-    selector: (row: any) => row.professor,
+    id: "periodo",
+    name: "Período",
+    selector: (row: any) => row.periodo,
     sortable: true,
+    width: "15%",
   },
   {
     id: "acoes",
     sortable: false,
     right: true,
     grow: 0,
-    cell: () => (
+    cell: (props: any) => (
       <div className="flex gap-2">
         <Excluir
-          title="Excluir Disciplina"
-          description="Tem certeza que deseja excluir esse disciplina?"
-          onClick={() => (console.log("Excluiu Disciplina!"))}
+          title="Excluir Curso"
+          description="Tem certeza que deseja excluir esse curso?"
+          onClick={() => (console.log("Excluiu curso!"))}
         />
         <button
           type="button"
@@ -60,28 +57,41 @@ const columns = [
     ),
   },
 ];
-
+/*
 const data = [
   {
     id: 1,
-    codigo: "GCC-218",
-    nome: "Engenharia da Computação",
+    nome: "Ciência da Computação",
     departamento: "Departamento de Ciência da Computação",
-    professor: "Antônio Maria",
     periodo: "2022/2",
   },
   {
     id: 2,
-    codigo: "GCC-244",
-    nome: "Práticas de Programação Orientada a Objetos",
+    nome: "Sistemas de Informação",
     departamento: "Departamento de Ciência da Computação",
-    professor: "Julio Cesar",
     periodo: "2022/2",
   },
 ];
+*/
+
 
 const Home: NextPage = () => {
   const [showCadastrar, setShowCadastrar] = useState(false);
+  const [cursos, setCursos] = useState([]);
+
+  useEffect(() => {
+    fetch("http://localhost:8080/api/curso/listar")
+      .then((response) => response.json())
+      .then((data) => {
+        data.map( (curso: any) => {
+          curso.departamento = curso.departamento.nome;
+        })
+        setCursos(data);
+      })
+      .catch((err) => {
+        console.log(err.message);
+      });
+  }, []);
 
   return (
     <Template
@@ -94,10 +104,11 @@ const Home: NextPage = () => {
         />
       }
     >
+      <Navbar links={NavbarAdminLinks} />
       <div className="container py-16">
         <div className="flex justify-between items-center">
-          <h2 className="text-gray-700">Disciplinas</h2>
-          <CadastrarDisciplina
+          <h2 className="text-gray-700">Cursos</h2>
+          <CadastrarCurso
             show={showCadastrar}
             setShow={setShowCadastrar}
           />
@@ -106,12 +117,12 @@ const Home: NextPage = () => {
         <div className="mt-8 overflow-x-auto animate-fade-in-up text-gray-700">
           <DataTable
             columns={columns}
-            data={data}
+            data={cursos}
             pagination
             paginationComponentOptions={paginationComponentOptions}
             highlightOnHover
             pointerOnHover
-            noDataComponent={<EmptyTable title="Não há disciplinas cadastrados :(" description="Cadastre um disciplina no botão Cadastrar!" icon={HiOutlineBookOpen} />}
+            noDataComponent={<EmptyTable title="Não há cursos cadastrados :(" description="Cadastre um curso no botão Cadastrar!" icon={HiFolderOpen} />}
           />
         </div>
       </div>
