+++ src/frontend/src/pages/admin/index.tsx
@@ -1,9 +1,9 @@
 import Card from "@components/cards/card";
-import { Cards } from "@utils/data";
+import Navbar from "@components/navigation/navbar";
+import { Cards, NavbarAdminLinks } from "@utils/data";
 import type { NextPage } from "next";
-import { useSession } from "next-auth/react";
-import { Meta } from "../templates/meta";
-import { Template } from "../templates/template";
+import { Meta } from "../../templates/meta";
+import { Template } from "../../templates/template";
 
 const Home: NextPage = () => {
   return (
@@ -17,6 +17,7 @@ const Home: NextPage = () => {
         />
       }
     >
+      <Navbar links={NavbarAdminLinks} />
       <div className="container py-16">
         <h1 className="text-gray-700 text-center">Seja Bem Vindo!</h1>
         <div className="mt-12 grid sm:grid-cols-2 lg:grid-cols-3 xl:grid-cols-5 gap-6">
