+++ src/frontend/src/styles/globals.css
@@ -43,6 +43,12 @@
 
 }
 
+@layer components {
+  .card-white {
+    @apply w-full bg-white rounded-2xl border border-gray-100 shadow-lg py-6 px-4;
+  }
+}
+
 .rdt_TableCol{
   font-size: 14px !important;
 }
