+++ src/frontend/src/templates/template.tsx
@@ -12,7 +12,6 @@ const Template = (props: ITemplateProps) => (
     {props.meta}
 
     <div className="pt-12 min-h-screen flex flex-col justify-between">
-      <Navbar />
       {props.children}
       <Footer />
     </div>
